Rename packages from processor to reduce
