+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/presentation/reducer/StartUpReducer.kt
@@ -1,8 +1,8 @@
-package com.gdavidpb.tuindice.login.presentation.processor
+package com.gdavidpb.tuindice.login.presentation.reducer
 
 import com.gdavidpb.tuindice.base.domain.model.StartUpAction
 import com.gdavidpb.tuindice.base.domain.usecase.base.UseCaseState
-import com.gdavidpb.tuindice.base.presentation.processor.BaseReducer
+import com.gdavidpb.tuindice.base.presentation.reducer.BaseReducer
 import com.gdavidpb.tuindice.login.domain.usecase.StartUpUseCase
 import com.gdavidpb.tuindice.login.domain.usecase.error.StartUpError
 import com.gdavidpb.tuindice.login.presentation.contract.Splash
